refactor: Move ChartAnnotator to different file + rename
